+++ src/main/java/com/example/spot/domain/LikedPost.java
@@ -1,7 +1,8 @@
-package com.example.domain;
+package com.example.spot.domain;
 
+import com.example.spot.domain.Member;
+import com.example.spot.domain.Post;
 import com.example.spot.domain.common.BaseEntity;
-import com.example.spot.domain.study.Study;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -10,17 +11,17 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RegionStudy extends BaseEntity {
-
+public class LikedPost extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "region_id")
-    private Region region;
+    @JoinColumn(name = "post_id")
+    private Post post;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "study_id")
-    private Study study;
+    @JoinColumn(name = "member_id")
+    private Member member;
+
 }
