+++ src/main/java/com/example/spot/domain/LikedPostComment.java
@@ -1,7 +1,7 @@
-package com.example.domain;
+package com.example.spot.domain;
 
 import com.example.spot.domain.Member;
-import com.example.spot.domain.Post;
+import com.example.spot.domain.PostComment;
 import com.example.spot.domain.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -11,14 +11,16 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LikedPost extends BaseEntity {
+public class LikedPostComment extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    private boolean is_liked; //좋아요:1, 싫어요:0
+
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
+    @JoinColumn(name = "post_comment_id")
+    private PostComment postComment;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
