+++ src/main/java/com/example/spot/domain/enums/Board.java
@@ -1,4 +1,4 @@
-package com.example.domain;
+package com.example.spot.domain.enums;
 
 public enum Board {
      합격후기, 정보공유, 고민상담, 취준토크, 자유토크, SPOT공지
