+++ src/main/java/com/example/spot/domain/enums/PostStatus.java
@@ -1,4 +1,4 @@
-package com.example.domain;
+package com.example.spot.domain.enums;
 
 public enum PostStatus {
     신고접수, 정상, 삭제
