+++ src/main/java/com/example/spot/domain/mapping/PreferredRegion.java
@@ -1,6 +1,6 @@
 package com.example.spot.domain.mapping;
 
-import com.example.domain.Region;
+import com.example.spot.domain.Region;
 import com.example.spot.domain.Member;
 import jakarta.persistence.*;
 import lombok.Getter;
