+++ src/main/java/com/example/spot/domain/mapping/RegionStudy.java
@@ -1,8 +1,8 @@
-package com.example.domain;
+package com.example.spot.domain.mapping;
 
-import com.example.spot.domain.Member;
-import com.example.spot.domain.Post;
+import com.example.spot.domain.Region;
 import com.example.spot.domain.common.BaseEntity;
+import com.example.spot.domain.study.Study;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -11,21 +11,17 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PostReport extends BaseEntity {
+public class RegionStudy extends BaseEntity {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String content;
-
-    @Enumerated(EnumType.STRING)
-    private PostStatus postStatus;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "region_id")
+    private Region region;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
+    @JoinColumn(name = "study_id")
+    private Study study;
 }
